refactor(api): split protocolStats.ts into focused modules
